+++ ui/pages/bridge/prepare/components/bridge-alert-modal.test.tsx
@@ -13,7 +13,7 @@ import { DummyQuotesNoApproval } from '../../../../../test/data/bridge/dummy-quo
 import { enLocale as messages } from '../../../../../test/lib/i18n-helpers';
 import configureStore from '../../../../store/store';
 import { HardwareWalletProvider } from '../../../../contexts/hardware-wallets/HardwareWalletContext';
-import * as useSubmitBridgeTransactionModule from '../../hooks/useSubmitBridgeTransaction';
+import * as useSubmitBridgeTransactionModule from '../../../../hooks/bridge/useSubmitBridgeTransaction';
 import * as bridgeSelectors from '../../../../ducks/bridge/selectors';
 import { BridgeAlert } from '../types';
 import { BridgeAlertModal } from './bridge-alert-modal';
@@ -21,7 +21,7 @@ import { BridgeAlertModal } from './bridge-alert-modal';
 const mockOnClose = jest.fn();
 const mockSubmitBridgeTransaction = jest.fn();
 
-jest.mock('../../hooks/useSubmitBridgeTransaction', () => ({
+jest.mock('../../../../hooks/bridge/useSubmitBridgeTransaction', () => ({
   // eslint-disable-next-line @typescript-eslint/naming-convention
   __esModule: true,
   default: jest.fn(() => ({
