+++ ui/pages/bridge/prepare/components/bridge-alert-modal.tsx
@@ -30,7 +30,7 @@ import {
   getValidationErrors,
 } from '../../../../ducks/bridge/selectors';
 import { Column, Row } from '../../layout';
-import useSubmitBridgeTransaction from '../../hooks/useSubmitBridgeTransaction';
+import useSubmitBridgeTransaction from '../../../../hooks/bridge/useSubmitBridgeTransaction';
 import { useBridgeAlerts } from '../../hooks/useBridgeAlerts';
 import { type BridgeAlert } from '../types';
